Move parallax snap sample to separate package
