+++ app/src/main/java/com/flaringapp/compose/topbar/MainActivity.kt
@@ -20,7 +20,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.flaringapp.compose.topbar.ui.samples.ParallaxSnapCollapsingSample
+import com.flaringapp.compose.topbar.ui.samples.advanced.ParallaxSnapCollapsingSample
 import com.flaringapp.compose.topbar.ui.theme.ComposeCollapsingTopBarTheme
 
 class MainActivity : ComponentActivity() {
