+++ app/src/main/java/com/flaringapp/compose/topbar/ui/samples/advanced/ParallaxSnapCollapsingSample.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.flaringapp.compose.topbar.ui.samples
+package com.flaringapp.compose.topbar.ui.samples.advanced
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Surface
